\name{meisowicz}
\alias{meisowicz}
\title{Meisowicz viscosities}
\description{Compute Meisowicz viscosities $\eta _1$, $\eta _2$, $\eta _3$
plus $\eta _{12}$, $\gamma _1$ and $\gamma _2$ from Leslie viscosities $\alpha _1, \ldots, \alpha _5$. Helfrech notation is used, 
where $\eta _1$ and $\eta _2$ are swapped relative
to Meisowicz's original work}
\usage{meisowicz(alpha)}
\arguments{
\item{alpha}{first five Leslie viscosities}
}
\seealso{}
\examples{}
\keyword{}
